(refactor): refactored the code in management team and advisory team into separate components
