+++ src/pages/Baobabpad.tsx
@@ -21,96 +21,12 @@ import { useState } from "react";
 import Internship from "@/components/Impact/Internship";
 import Mentorship from "@/components/Impact/Mentorship";
 import LinkedIn from "next-auth/providers/linkedin";
+import Management from "@/components/baobabpad_sections/Management";
+import AdvisoryTeam from "@/components/baobabpad_sections/AdvisoryTeam";
 
 const Baobabpad = () => {
-  const teamDetailsArray = [
-    {
-      image: "/brenda_mkeshwa_CEO.jpg",
-      username: "Brenda Mkeshwa",
-      position: "CEO",
-    },
-    {
-      image: "/kenny.png",
-      username: "Kenny Kaluiji",
-      position: "Founder",
-    },
-    {
-      image: "/jacob_mwale_tech_lead.jpg",
-      username: "Jacob Mwale",
-      position: "Tech Lead",
-    },
-    {
-      image: "/chimwemwe_masona_tech_lead.png",
-      username: "Chimwemwe Masona",
-      position: "Tech Lead",
-    },
-    {
-      image: "/selma_hamutenya_product_manager.png",
-      username: "Selma Hamutenya",
-      position: "Product Manager",
-    },
-    {
-      image: "/Mark_Ai.jpg",
-      username: "Mark Njoroge",
-      position: "AI Lead",
-    },
-    {
-      image: "/steve.jpg",
-      username: "Stephen Ondieki",
-      position: "Testing Lead",
-    },
-    {
-      image: "/crispin_manda_product_manager.png",
-      username: "Crispin Manda ",
-      position: "Product Manager",
-    },
-    {
-      image: "/sidney_kipron_cloud_manager.png",
-      username: "Sidney Kiprono ",
-      position: "Cloud Manager",
-    },
-    {
-      image: "/martin_mundia_cloud_manager.jpg",
-      username: "Martin Mundia ",
-      position: "Cloud Manager",
-    },
-  ];
+ 
 
-  const adversoryTeamArray = [
-    {
-      image: "/Kofo Obasanjo.jpeg",
-      username: "Kofo Blackshire",
-      position: "Changemaker",
-      LinkedIn:
-        "https://www.linkedin.com/in/kofo-obasanjo-blackshire-80a302a4/",
-    },
-    {
-      image: "/Henry Ogbuagu.jpeg",
-      username: "Henry Ogbuagu",
-      position: "Business Advisor",
-      LinkedIn: "https://www.linkedin.com/in/henry-ogbuagu-b1156283/",
-    },
-
-    {
-      image: "/Lamisi Gurah.jpeg",
-      username: "Lamisi Gurah",
-      position: "Femtech Advisor",
-      LinkedIn: "https://www.linkedin.com/in/lamisi/",
-    },
-
-    {
-      image: "/Yannick_Di_Mondo.jpg",
-      username: "Yannick di Mondo",
-      position: "Technology Advisor",
-      LinkedIn: "https://www.linkedin.com/in/yannickdimondo/",
-    },
-    {
-      image: "/Tish_Gilberts.jpeg",
-      username: "Tish Gilbert",
-      position: "Impact Advisor",
-      LinkedIn: "https://www.linkedin.com/in/tish-gilbert-2b084497/",
-    },
-  ];
 
   const [showIntern, setShowIntern] = useState(false);
   const toggleInternship = () => {
@@ -121,15 +37,7 @@ const Baobabpad = () => {
     setShowMentor(!showMentor);
   };
 
-  const [showAllCards, setShowAllCards] = useState(false);
-  const visibleCards = showAllCards
-    ? teamDetailsArray
-    : teamDetailsArray.slice(0, 2);
-  const buttonLabel = showAllCards ? "See Less" : "See More";
 
-  const toggleShowAllCards = () => {
-    setShowAllCards(!showAllCards);
-  };
 
   return (
     <div className="min-h-screen flex flex-col">
@@ -230,119 +138,10 @@ const Baobabpad = () => {
         {/* end of section */}
 
         {/* Start of  Village Management Team  */}
-
-        <section className="border-t border-500 p-12">
-          <div className=" p-2 flex flex-col gap-y-1 justify-center">
-            <span className="font-bold text-600 text-[1.5rem]">
-              Village Management Team
-            </span>
-
-            {/*  start of profile cards view for small screen */}
-            <div className=" px-4  sm:flex gap-x-6   sm:flex-wrap justify-center md:hidden">
-              {visibleCards.map((member, index) => (
-                <div
-                  key={index}
-                  className="pt-2 w-[15rem] h-[20rem] flex flex-col items-center justify-center gap-y-4 my-12 border border-600  shadow-2xl hover:transform hover:scale-110 transition-transform duration-300 hover:cursor-pointer
-                      "
-                >
-                  <img
-                    src={member.image}
-                    alt="Team Member"
-                    className="w-[40%] h-[30%] rounded-[50%]"
-                  />
-                  <span className="font-bold font-700 text-xl">
-                    {member.username}
-                  </span>
-                  <span className="text-[#9095A0]">{member.position}</span>
-                  <button
-                    type="button"
-                    className="bg-[#C8F9FF] pt-2 pb-2 px-8 py-8 align-center "
-                  >
-                    View Profile
-                  </button>
-                  {/* <div className="bg-[#C8F9FF] text-white rounded-[100%] w-10 h-10 px-3 py-1  "><span className="text-black  text-2xl">+</span></div> */}
-                </div>
-              ))}
-            </div>
-            {/* end of profile cards view for small screen*/}
-            <button
-              onClick={() => setShowAllCards(!showAllCards)}
-              className={`mt-4 bg-blue-500 hover:bg-blue-700 text-white font-bold py-2 px-4 rounded sm:block   ${
-                showAllCards ? "" : "md:hidden"
-              }`}
-            >
-              {buttonLabel}
-            </button>
-              
-
-              {/* Medium screen and above view for the team profile */}
-             <div className="px-4 md:flex gap-x-6 md:flex-wrap md:justify-center hidden">
-              {teamDetailsArray.map((member, index) => (
-                <div
-                  key={index}
-                  className="pt-2 w-[15rem] h-[20rem] flex flex-col items-center justify-center gap-y-4 my-12 border border-600 shadow-2xl hover:transform hover:scale-110 transition-transform duration-300 hover:cursor-pointer"
-                >
-                  <img
-                    src={member.image}
-                    alt="Team Member"
-                    className="w-[40%] h-[30%] rounded-[50%]"
-                  />
-                  <span className="font-bold font-700 text-xl">
-                    {member.username}
-                  </span>
-                  <span className="text-[#9095A0]">{member.position}</span>
-                  <button
-                    type="button"
-                    className="bg-[#C8F9FF] pt-2 pb-2 px-8 py-8 align-center"
-                  >
-                    View Profile
-                  </button>
-                </div>
-              ))}
-             </div>
-            {/* end of the medium and above view for the team profile */}
-
-          </div>
-        </section>
-        <section className="border-t border-500 p-12">
-          <div className=" p-4 flex flex-col gap-y-1 justify-center">
-            <span className="font-bold text-600 text-[1.5rem]">
-              Advisory Team
-            </span>
-
-            {/*  start of profile cards */}
-            <div className=" px-4  flex gap-x-6  flex-wrap justify-center">
-              {adversoryTeamArray.map((member, index) => (
-                <div
-                  key={index}
-                  className="pt-2 w-[15rem] h-[20rem] flex flex-col items-center justify-center gap-y-4 my-12 border border-600  shadow-2xl hover:transform hover:scale-110 transition-transform duration-300 hover:cursor-pointer "
-                >
-                  <img
-                    src={member.image}
-                    alt="Team Member"
-                    className="w-[40%] h-[30%] rounded-[50%]"
-                  />
-                  <span className="font-bold font-700 text-xl">
-                    {member.username}
-                  </span>
-                  <span className="text-[#9095A0]">{member.position}</span>
-                  <div className=" pt-2 pb-2 px-8 py-8   flex flex-row ">
-                    <img src="/linkedIn.png" alt="" className="px-2" />
-                    <a
-                      href={member.LinkedIn}
-                      className="hover:text-[#00BDD6] hover:text-underline"
-                      target="_blank"
-                    >
-                      View Profile
-                    </a>
-                  </div>
-                  {/* <div className="bg-[#C8F9FF] text-white rounded-[100%] w-10 h-10 px-3 py-1  "><span className="text-black  text-2xl">+</span></div> */}
-                </div>
-              ))}
-            </div>
-            {/* end of profile cards */}
-          </div>
-        </section>
+           <Management/>
+           <AdvisoryTeam/>
+        
+        
         {/* end of Village management team  */}
 
         {/* Start of why baobabpad section */}
